+++ sentry/src/test/java/io/sentry/clientreport/ClientReportRecorderTest.kt
@@ -1,0 +1,3 @@
+package io.sentry.clientreport
+
+public class ClientReportRecorderTest {}